# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Savings Account API
  description: API for managing savings accounts.
  version: 1.0.0
paths:
  /accounts:
    post:
      tags:
        - Savings Accounts
      summary: Create a new savings account for a consumer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Validation errors
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /accounts/{id}:
    get:
      tags:
        - Savings Accounts
      summary: Get savings account details by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    CreateAccountRequest:
      type: object
      required:
        - customerName
        - accountNickName
      properties:
        customerName:
          type: string
          minLength: 1
          maxLength: 255
        accountNickName:
          type: string
          minLength: 5
          maxLength: 30
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerName:
          type: string
        accountNumber:
          type: string
        accountNickName:
          type: string
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string